allow putting list elements in configuration
These allow the scoreboard to randomly pick an index of possible lists
in the configuration to allow checking more diverse things per service.
Usage is to simply replace options desired to change all with lists of
the same size for a given service. One index will be picked for all
lists in the service.